Extract enums into separate headers
