+++ src/SloppinessEnum.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+enum Sloppiness {
+  SLOPPY_INCLUDE_FILE_MTIME = 1 << 0,
+  SLOPPY_INCLUDE_FILE_CTIME = 1 << 1,
+  SLOPPY_TIME_MACROS = 1 << 2,
+  SLOPPY_PCH_DEFINES = 1 << 3,
+  // Allow us to match files based on their stats (size, mtime, ctime), without
+  // looking at their contents.
+  SLOPPY_FILE_STAT_MATCHES = 1 << 4,
+  // Allow us to not include any system headers in the manifest include files,
+  // similar to -MM versus -M for dependencies.
+  SLOPPY_SYSTEM_HEADERS = 1 << 5,
+  // Allow us to ignore ctimes when comparing file stats, so we can fake mtimes
+  // if we want to (it is much harder to fake ctimes, requires changing clock)
+  SLOPPY_FILE_STAT_MATCHES_CTIME = 1 << 6,
+  // Allow us to not include the -index-store-path option in the manifest hash.
+  SLOPPY_CLANG_INDEX_STORE = 1 << 7,
+  // Ignore locale settings.
+  SLOPPY_LOCALE = 1 << 8,
+  // Allow caching even if -fmodules is used.
+  SLOPPY_MODULES = 1 << 9,
+};
