+++ src/argprocessing.hpp
@@ -19,7 +19,6 @@
 #pragma once
 
 #include "Args.hpp"
-#include "Statistics.hpp"
 
 #include "third_party/nonstd/optional.hpp"
