+++ src/exceptions.hpp
@@ -21,7 +21,7 @@
 #include "system.hpp"
 
 #include "FormatNonstdStringView.hpp"
-#include "Statistics.hpp"
+#include "StatisticEnum.hpp"
 
 #include "third_party/fmt/core.h"
 #include "third_party/nonstd/optional.hpp"
