+++ unittest/test_Config.cpp
@@ -17,8 +17,8 @@
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
 #include "../src/Config.hpp"
+#include "../src/SloppinessEnum.hpp"
 #include "../src/Util.hpp"
-#include "../src/ccache.hpp"
 #include "../src/exceptions.hpp"
 #include "../src/fmtmacros.hpp"
 #include "TestUtil.hpp"
